+++ narrator.ts
@@ -62,14 +62,17 @@ const NarratorTools = {
 	_element: $(
 		'<div id="narrator" class="narrator"><div class="narrator-bg"></div><div class="narrator-frame"><div class="narrator-frameBG"></div><div class="narrator-box"><div class="narrator-content"></div></div><div class="narrator-buttons" style="opacity:0;"><button class="NT-btn-pause"></button><button class="NT-btn-close"></button></div></div><div class="narrator-sidebarBG"></div>'
 	),
-	_character: '',
+	/**
+	 * Here is where a custom speaker is stored, if you change the value to anything other than '' the next messages will speak as such
+	 */
+	character: '',
 	/**
 	 * Hooked function wich identifies if a message is a Narrator Tools command
-	 * @param _message
+	 * @param message
 	 * @param content   Message to be identified
 	 * @param chatData
 	 */
-	_chatMessage(_message: string, content: string, _chatData: any) {
+	_chatMessage(message: string, content: string, chatData: any) {
 		if (!game.user.isGM) return;
 		const narration = new RegExp('^\\/narrat(?:e|ion) ([^]*)', 'i');
 		const description = new RegExp('^\\/desc(?:ribe|ription|) ([^]*)', 'i');
@@ -78,7 +81,7 @@ const NarratorTools = {
 		const commands = {
 			narration: narration,
 			description: description,
-			note: notification,
+			notification: notification,
 			as: as,
 		};
 		// Iterate over patterns, finding the first match
@@ -88,10 +91,10 @@ const NarratorTools = {
 			if (match) {
 				if (c == 'as') {
 					if (match[1]) {
-						this._character = match[1];
-						($('#chat-message')[0] as HTMLInputElement).placeholder = game.i18n.localize('NT.SpeakingAs') + ' ' + this._character;
+						this.character = match[1];
+						($('#chat-message')[0] as HTMLInputElement).placeholder = game.i18n.localize('NT.SpeakingAs') + ' ' + this.character;
 					} else {
-						this._character = '';
+						this.character = '';
 						($('#chat-message')[0] as HTMLInputElement).placeholder = '';
 					}
 				} else {
@@ -124,7 +127,10 @@ const NarratorTools = {
 		/**If a narration had ocurred and the display now is still on, turn it off */
 		if (!narration.display && this.elements.content[0].style.opacity === '1') {
 			this.elements.BG.height(0);
-			if (game.user.isGM) this.elements.buttons[0].style.opacity = '0';
+			if (game.user.isGM) {
+				this.elements.buttons[0].style.opacity = '0';
+				this.elements.buttons[0].style.visibility = 'hidden';
+			}
 		}
 
 		/**If the message suddenly disappears, turn off the opacity */
@@ -165,7 +171,7 @@ const NarratorTools = {
 				}
 			};
 
-			/**If the display is on and the narration.id is a new one, it means a new narration is taking place */
+			/** If the display is on and the narration.id is a new one, it means a new narration is taking place */
 			if (narration.id !== this._id) {
 				this._id = narration.id;
 				clearTimeout(this._timeouts.narrationOpens);
@@ -182,6 +188,7 @@ const NarratorTools = {
 
 					if (game.user.isGM) {
 						this.elements.buttons[0].style.opacity = '1';
+						this.elements.buttons[0].style.visibility = 'visible';
 						this.elements.buttons[0].style.top = `calc(50% + ${60 + height / 2}px)`;
 						this._updateStopButton(game.settings.get('narrator-tools', 'NarrationStartPaused'));
 					}
@@ -192,7 +199,7 @@ const NarratorTools = {
 
 				Hooks.once('narration', scroll);
 			} else {
-				/**If narration is paused, stop animation and clear timeouts */
+				/** If narration is paused, stop animation and clear timeouts */
 				if (narration.paused) {
 					if (this._timeouts.narrationScrolls) {
 						clearTimeout(this._timeouts.narrationScrolls);
@@ -234,9 +241,9 @@ const NarratorTools = {
 	},
 	/**
 	 * Hides the journals context menu
-	 * @param _e
+	 * @param e
 	 */
-	_hideContextMenu(_e: Event) {
+	_hideContextMenu(e: Event) {
 		NarratorTools._menu.hide();
 		document.removeEventListener('click', NarratorTools._hideContextMenu);
 	},
@@ -275,15 +282,15 @@ const NarratorTools = {
 		}
 	},
 	/**
-	 * Creates an alias and change message type if this._character option is true
+	 * Creates an alias and change message type if this.character option is true
 	 * @param chatData Change the chat message configuration
-	 * @param _options
-	 * @param _user
+	 * @param options
+	 * @param user
 	 */
-	_preCreateChatMessage(chatData: any, _options: any, _user: string) {
-		if (game.user.isGM && this._character) {
+	_preCreateChatMessage(chatData: any, options: any, user: string) {
+		if (game.user.isGM && this.character) {
 			chatData.type = this._msgtype;
-			chatData.speaker = { alias: this._character };
+			chatData.speaker = { alias: this.character };
 		}
 	},
 	/**Initialization routine for 'ready' hook */
@@ -457,7 +464,7 @@ const NarratorTools = {
 	/**Specify how the module's messages will be intepreted by foundry and other modules:
 	 * OTHER: 0, OOC: 1, IC: 2, EMOTE: 3, WHISPER: 4, ROLL: 5
 	 */
-	_msgtype: CONST.CHAT_MESSAGE_TYPES.OTHER,
+	_msgtype: CONST.CHAT_MESSAGE_TYPES.IC,
 	/**
 	 * Behavior when a chat message is clicked
 	 * @param event The event wich triggered the handler
@@ -487,29 +494,30 @@ const NarratorTools = {
 	 * Renders the chat message and sets out the message behavior
 	 * @param message Message object to be rendered
 	 * @param html HTML element of the message
-	 * @param _data
+	 * @param data
 	 */
-	_renderChatMessage(message: any, html: JQuery<HTMLElement>, _data: any) {
-		const span = html.find('.narrator-span');
-		if (span.length) {
+	_renderChatMessage(message: any, html: JQuery<HTMLElement>, data: any) {
+		const type = message.getFlag('narrator-tools', 'type');
+		if (type) {
+			html.find('div.message-content')[0].innerText = html.find('div.message-content')[0].innerText.slice(9);
 			html.find('.message-sender').text('');
 			html.find('.message-metadata')[0].style.display = 'none';
 			html[0].classList.add('narrator-chat');
-			if (span[0].classList.contains('narration')) {
+			if (type == 'narration') {
 				html[0].classList.add('narrator-narrative');
-			} else if (span[0].classList.contains('description')) {
+			} else if (type == 'description') {
 				html[0].classList.add('narrator-description');
-			} else if (span[0].classList.contains('note')) {
+			} else if (type == 'notification') {
 				html[0].classList.add('narrator-notification');
 			}
 		}
 	},
 	/**
 	 * Hook wich triggers when the journal sheet is rendered
-	 * @param _journalSheet
+	 * @param journalSheet
 	 * @param html
 	 */
-	_renderJournalSheet(_journalSheet: any, html: JQuery<HTMLElement>) {
+	_renderJournalSheet(journalSheet: any, html: JQuery<HTMLElement>) {
 		let editor = '.editor-content';
 		// Identifies if there is a Easy MDE Container
 		const MDEContainer = html.find('.EasyMDEContainer').length;
@@ -621,7 +629,7 @@ const NarratorTools = {
 		 * @param options - Change the chat message configuration
 		 */
 		notify(message: string, options = {}) {
-			return NarratorTools.createChatMessage('note', message, options);
+			return NarratorTools.createChatMessage('notification', message, options);
 		},
 	},
 	/**
@@ -636,7 +644,12 @@ const NarratorTools = {
 		message = message.replace(/\\n|<br>/g, '\n');
 
 		let chatData = {
-			content: `<span class="narrator-span ${type}">${message}</span>`,
+			content: message,
+			flags: {
+				'narrator-tools': {
+					type: type,
+				},
+			},
 			type: this._msgtype,
 			speaker: {
 				alias: game.i18n.localize('NT.Narrator'),